+++ src/client/src/api-lib/index.js
@@ -2,7 +2,7 @@ import { getHeight } from "./actions";
 import { deleteBoreholes, patchBoreholes } from "./actions/borehole";
 import { loadSettings, patchSettings } from "./actions/settings";
 import { draftTerms, getTerms, getTermsDraft, publishTerms } from "./actions/terms";
-import { loadUser, setAuthentication, unsetAuthentication } from "./actions/user";
+import { setAuthentication, unsetAuthentication } from "./actions/user";
 import store, { injectReducer } from "./reducers";
 
 export {
@@ -15,7 +15,6 @@ export {
   publishTerms,
   setAuthentication,
   unsetAuthentication,
-  loadUser,
   deleteBoreholes,
   patchBoreholes,