+++ src/client/src/auth/AuthOverlay.tsx
@@ -1,19 +1,9 @@
-import { FC, PropsWithChildren, useEffect } from "react";
+import { FC, PropsWithChildren } from "react";
 import { useTranslation } from "react-i18next";
-import { useDispatch, useSelector } from "react-redux";
 import { Alert, Button, CircularProgress } from "@mui/material";
-import { loadUser } from "../api-lib";
+import { useCurrentUser } from "../api/user.ts";
 import { SplashScreen } from "./SplashScreen.tsx";
 import { useAuth } from "./useBoreholesAuth.tsx";
-
-interface User {
-  data: object;
-  error?: string;
-}
-
-interface ReduxState {
-  core_user: User;
-}
 
 export const AuthOverlay: FC<PropsWithChildren> = ({ children }) => {
   const auth = useAuth();
@@ -23,9 +13,8 @@ export const AuthOverlay: FC<PropsWithChildren> = ({ children }) => {
     auth.isAuthenticated = true;
   }
 
-  const dispatch = useDispatch();
-  const user = useSelector<ReduxState, User>(state => state.core_user);
   const { t } = useTranslation();
+  const { data: user, isError } = useCurrentUser(auth.isAuthenticated);
 
   const signIn = () => {
     const path = `${window.location.pathname}${window.location.search}${window.location.hash}`;
@@ -38,19 +27,12 @@ export const AuthOverlay: FC<PropsWithChildren> = ({ children }) => {
     auth.signoutRedirect();
   };
 
-  useEffect(() => {
-    if (auth.isAuthenticated) {
-      // @ts-expect-error legacy API methods will not be typed, as they are going to be removed
-      dispatch(loadUser());
-    }
-  }, [auth.isAuthenticated, dispatch]);
-
-  if (auth.isAuthenticated && user.data) {
+  if (auth.isAuthenticated && user) {
     return children;
   } else if (!auth.isLoading && !auth.isAuthenticated) {
     // Perform automatic login if user is not authenticated.
     signIn();
-  } else if (user?.error) {
+  } else if (isError) {
     return (
       <SplashScreen>
         <Alert severity="error">{t("userUnauthorized")}</Alert>
@@ -59,11 +41,10 @@ export const AuthOverlay: FC<PropsWithChildren> = ({ children }) => {
         </Button>
       </SplashScreen>
     );
-  } else {
-    return (
-      <SplashScreen>
-        <CircularProgress />
-      </SplashScreen>
-    );
   }
+  return (
+    <SplashScreen>
+      <CircularProgress />
+    </SplashScreen>
+  );
 };